// web/src/app/_api/metrics-api.ts
export enum MetricNames {
  Temperature = "temperature",
  Humidity = "humidity",
}

export interface MetricModel {
  /** @format date-time */
  timestamp: string;
  name: MetricNames;
  /** @format double */
  value: number;
}